track-software/py-proj/requirements.txt
future==1.0.0
iso8601==2.1.0
linkify-it-py==2.0.3
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
Pygments==2.17.2
pyserial==3.5
python-i18n==0.3.9
PyYAML==6.0.1
readchar==4.0.5
rich==13.7.0
typing_extensions==4.9.0
uc-micro-py==1.0.3
wcwidth==0.2.13
